Update tic_tac_toe.c
this is a better version of tic_tac_toe.
here computer will calculate the no. of wins and depth differently and check them differently.
First it saves all the move with max wins and takes those move among them where the depth is minimun and chooses the ultimate move among them randomly.